bloom is uint64.

+++ libdariadb/storage/bloom_filter.h
@@ -6,17 +6,17 @@
 namespace dariadb {
 namespace storage {
 
-template <typename T> static size_t bloom_empty() {
+template <typename T> static uint64_t bloom_empty() {
   return size_t();
 }
 
-template <typename T> static size_t bloom_add(const size_t &fltr, const T &val) {
-  auto h = std::hash<T>()(val);
+template <typename T> static uint64_t bloom_add(const uint64_t &fltr, const T &val) {
+  auto h = static_cast<uint64_t>(std::hash<T>()(val));
   return fltr | h;
 }
 
-template <typename T> static bool bloom_check(const size_t &fltr, const T &val) {
-  auto h = std::hash<T>()(val);
+template <typename T> static bool bloom_check(const uint64_t &fltr, const T &val) {
+  auto h = static_cast<uint64_t>(std::hash<T>()(val));
   return (fltr & h) == h;
 }
 }
